Create 106. Construct Binary Tree from Inorder and Postorder Traversal
There is a string thing for Python. If the parameter is an array. recursively pop one element  and the original array is modified.  Below is the test.

def test(array, times, cluster):
    #print(array)
    if array:
        print('times is %d, cluster is %d' %(times, cluster))
        if times > 2:
            return
        print('array is', array)
        array.pop()
        test(array, times+1, 1)
        test(array, times+2, 2)

array=[1,2,3,4,5]
test(array,0, 0)

Result is:
times is 0, cluster is 0
array is [1, 2, 3, 4, 5]


times is 1, cluster is 1
array is [1, 2, 3, 4]

times is 2, cluster is 1
array is [1, 2, 3]

times is 3, cluster is 1


times is 4, cluster is 2


times is 3, cluster is 2

times is 2, cluster is 2
array is [1, 2]


times is 3, cluster is 1
times is 4, cluster is 2

Process finished with exit code 0