Rename 96. Unique Binary Search Trees to 96-1. Unique Binary Search Trees
After checking the question. When n = 3, we can find that there are three value can be root.  root = 1, 2, 3.
when root = 1, all other nodes are in the right subtree. So, left subtree have 0 node and right subtree have 2 nodes. Node as f(0)*f(2)
when root =2, it must be 1 as left tree and 3 as right tree. So, left subtee have 1 node and right subtree have 1 nodes. Note as f(1)*f(1)
when root =3, it must be all other nodes are in the right subtree. So, left subtree have 2 nodes and right subtree have 1 nodes. Note as f(2)*f(0)

So f(3)=  f(0)*f(2) + f(1)*f(1)+f(2)*f(0)

So this is a Dynamic programming issue. f(0)=1 as array is none and only one BST(None tree)
f(1) is only one node and one BST.

   1         3     3      2      1
    \       /     /      / \      \
     3     2     1      1   3      2
    /     /       \                 \
   2     1         2                 3